+++ Digital Pantry/Model/Preferences.swift
@@ -10,30 +10,30 @@ import Foundation
 
 struct Preference {
     //properties/members of Preference
-    var preferenceID: Int
+    var preferenceID: Int64
     var type: String
-    var typeID: Int
+    var typeID: Int64
     
     //initialised where we have not obtained a UniqueIdentifier (preferenceID) as yet
-    init?(type: String, typeID: Int) {
+    init?(type: String, typeID: Int64) {
         self.preferenceID = -1
         self.type = type
         self.typeID = typeID
     }
     
     //use case: calling data back from DB/table
-    init?(preferenceID: Int, type: String, typeID: Int ) {
+    init?(preferenceID: Int64, type: String, typeID: Int64 ) {
         self.preferenceID = preferenceID
         self.type = type
         self.typeID = typeID
         
     }
     
-    mutating func setDietCategoryID(preferenceID: Int) { //once the DB has a preferenceID, use this to set it
+    mutating func setDietCategoryID(preferenceID: Int64) { //once the DB has a preferenceID, use this to set it
         self.preferenceID = preferenceID
     }
     
-    func getPreferenceID() -> Int { //retrieve preferenceID
+    func getPreferenceID() -> Int64 { //retrieve preferenceID
         return preferenceID
     }
         
@@ -45,11 +45,11 @@ struct Preference {
         self.type = type
     }
 
-    func getTypeID() -> Int { //retrieve typeID
+    func getTypeID() -> Int64 { //retrieve typeID
         return typeID
     }
     
-    mutating func setTypeID(typeID: Int) { //add a typeID to item
+    mutating func setTypeID(typeID: Int64) { //add a typeID to item
         self.typeID = typeID
     }
 }
